requirements.txt
# EditalShield Dependencies

# Core
click>=8.1
jinja2>=3.0
pydantic>=2.0
python-dotenv>=1.0
requests>=2.28

# Data & ML
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9

# Visualization
matplotlib>=3.8
seaborn>=0.13

# CLI Interface
rich>=13.0

# Web Scraping
beautifulsoup4>=4.12
lxml>=4.9

# Development
jupyter>=1.0
pytest>=7.0
black>=23.0
